CloudFormation descriptor that references managed secrets.
