+++ deployments/src/main/scala/net/wayfarerx/loremaster/deployments/Messages.scala
@@ -1,0 +1,34 @@
+/* Messages.scala
+ *
+ * Copyright (c) 2022 wayfarerx (@thewayfarerx).
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0.
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package net.wayfarerx.loremaster
+package deployments
+
+/**
+ * The messages provided by the deploy package.
+ */
+private object Messages {
+
+  def okay: String = "200 OK"
+
+  def accountId: String = s"The ID of the AWS account $Application operates in."
+
+  def failedToAccessConfiguration(key: String, thrown: Throwable): String =
+    s"""Failed to access configuration "$key": ${describe(thrown)}"""
+
+  def failedToDeliverSqsMessage(message: String): String =
+    s"Failed to deliver SQS message: $message"
+
+  def failedToParseSqsMessage(message: String): String =
+    s"Failed to parse SQS message: $message"
+
+}
