+++ project/plugins.sbt
@@ -1,3 +1,5 @@
+addSbtPlugin("com.github.sbt" % "sbt-proguard" % "0.5.0")
+
 addSbtPlugin("com.dwolla.sbt" % "sbt-s3-publisher" % "1.3.2")
 
-addSbtPlugin("com.github.sbt" % "sbt-proguard" % "0.5.0")
+addSbtPlugin("com.dwolla.sbt" % "sbt-cloudformation-stack" % "1.3.1")
